+++ src/ajax/ajax.js
@@ -26,16 +26,16 @@ KISSY.add('ajax', function(S, undef) {
             data: null,
             xhr: win.XMLHttpRequest ?
                 function() {
-                    return new window.XMLHttpRequest();
+                    return new win.XMLHttpRequest();
                 } :
                 function() {
                     try {
-                        //return new window.ActiveXObject('Microsoft.XMLHTTP');
-                        return S.UA.ie == 6 ? 
-								//iE 6 下请求缓存中的资源无法正确返回xhr.responseText,需要使用老版本的XMLHTTP
-								//jQuery 没有考虑请求缓存的情况
-								new window.ActiveXObject('Msxml2.XMLHTTP.5.0')://xhr 无状态值1
-								new window.ActiveXObject('Microsoft.XMLHTTP');
+                        // ie 6 下请求缓存中的资源无法正确返回 xhr.responseText, 需要使用老版本的 XMLHTTP
+                        // jQuery 没有考虑请求缓存的情况
+                        return new win.ActiveXObject(
+                            S.UA.ie == 6 ?
+                                'Msxml2.XMLHTTP.5.0' :
+                                'Microsoft.XMLHTTP');
                     } catch(e) {
                     }
                 },
@@ -117,7 +117,6 @@ KISSY.add('ajax', function(S, undef) {
                 xhr.setRequestHeader(CONTENT_TYPE, c.contentType);
             }
 
-            // Set the Accepts header for the server, depending on the dataType
             // Set the Accepts header for the server, depending on the dataType
 			xhr.setRequestHeader('Accept', c.dataType && c.accepts[c.dataType] ?
 				c.accepts[c.dataType] + ', */*; q=0.01' :
@@ -222,7 +221,7 @@ KISSY.add('ajax', function(S, undef) {
         jsonp: function(url, data, callback) {
             if(S.isFunction(data)) {
                 callback = data;
-				data = null;//占位符
+				data = null; // 占位符
             }
             return io.get(url, data, callback, JSONP);
         }
@@ -234,17 +233,17 @@ KISSY.add('ajax', function(S, undef) {
     S.jsonp = io.jsonp;
     S.IO = io;
     // 所有方法在 IO 下都可调 IO.ajax/get/post/getScript/jsonp
-    // S 下有便捷入口 S.io/S.ajax/getScript/jsonp
+    // S 下有便捷入口 S.io/ajax/getScript/jsonp
 
     //检测 xhr 是否成功
     function xhrSuccessful(xhr) {
         try {
 			// IE error sometimes returns 1223 when it should be 204 so treat it as success, see #1450
             // ref: http://www.w3.org/Protocols/rfc2616/rfc2616-sec10.html
-			// IE 中如果请求一个缓存住的页面，会出现如下状况 (jQuery中未考虑,此处也不作处理)：
-			// 		请求一个页面成功，但头输出为404,ie6/8下检测为200,ie7/ff/chrome/opera检测为404
-			// 		请求一个不存在的页面，ie 均检测为 200 ,ff/chrome/opera检测为404
-			// 		请求一个不存在的页面，ie6/7的statusText为'Not Found'，ie8的为'OK',statusText是可以被程序赋值的
+			// IE 中如果请求一个缓存住的页面，会出现如下状况 (jQuery 中未考虑,此处也不作处理)：
+			// 		请求一个页面成功，但头输出为 404, ie6/8 下检测为 200, ie7/ff/chrome/opera 检测为 404
+			// 		请求一个不存在的页面，ie 均检测为 200 ,ff/chrome/opera检测为 404
+			// 		请求一个不存在的页面，ie6/7 的 statusText为 'Not Found'，ie8 的为 'OK', statusText 是可以被程序赋值的
 			return xhr.status >= 200 && xhr.status < 300 ||
 				xhr.status === 304 || xhr.status === 1223;
 		} catch(e) {}