+++ src/dd/draggable.js
@@ -1,10 +1,10 @@
 /**
  * dd support for kissy, drag for dd
- * @author: yiminghe@gmail.com
+ * @author: 承玉<yiminghe@gmail.com>
  */
 KISSY.add('dd-draggable', function(S) {
 
-    var UA = S.UA,Node = S.Node;
+    var UA = S.UA;
 
     /*
      拖放纯功能类
@@ -20,19 +20,26 @@ KISSY.add('dd-draggable', function(S) {
          */
         node: {
             setter:function(v) {
-                return new Node(v);
+                return S.one(v);
             }
         },
 
         /**
-         * handler 集合，注意暂时必须在 node 里面
+         * 是否需要遮罩跨越iframe
+         */
+        shim:{
+            value:true
+        },
+
+        /**
+         * handler 数组，注意暂时必须在 node 里面
          */
         handlers:{
-            value:{},
+            value:[],
             setter:function(vs) {
                 if (vs) {
                     for (var i = 0; i < vs.length; i++) {
-                        vs[i] = new Node(vs[i]);
+                        vs[i] = S.one(vs[i]);
                     }
                 }
             }
@@ -46,19 +53,19 @@ KISSY.add('dd-draggable', function(S) {
                 node = self.get('node'),
                 handlers = self.get('handlers');
 
-            if (S.isEmptyObject(handlers)) {
-                handlers[node[0].id] = node;
+            if (handlers.length == 0) {
+                handlers[0] = node;
             }
 
-            for (var h in handlers) {
-                if (!handlers.hasOwnProperty(h)) continue;
-                var hl = handlers[h],
+            for (var i = 0; i < handlers.length; i++) {
+                var hl = handlers[i],
                     ori = hl.css('cursor');
                 if (hl[0] != node[0]) {
                     if (!ori || ori === 'auto')
                         hl.css('cursor', 'move');
                 }
             }
+
             node.on('mousedown', self._handleMouseDown, self);
         },
 
@@ -66,9 +73,8 @@ KISSY.add('dd-draggable', function(S) {
             var self = this,
                 node = self.get('node'),
                 handlers = self.get('handlers');
-            for (var h in handlers) {
-                if (!handlers.hasOwnProperty(h)) continue;
-                var hl = handlers[h];
+            for (var i = 0; i < handlers.length; i++) {
+                var hl = handlers[i];
                 if (hl.css("cursor") == "move") {
                     hl.css("cursor", "auto");
                 }
@@ -79,12 +85,12 @@ KISSY.add('dd-draggable', function(S) {
         _check: function(t) {
             var handlers = this.get('handlers');
 
-            for (var h in handlers) {
-                if (!handlers.hasOwnProperty(h)) continue;
-                if (handlers[h].contains(t)
+            for (var i = 0; i < handlers.length; i++) {
+                var hl = handlers[i];
+                if (hl.contains(t)
                     ||
                     //子区域内点击也可以启动
-                    handlers[h][0] == t[0]) return true;
+                    hl[0] == t[0]) return true;
             }
             return false;
         },
@@ -98,6 +104,7 @@ KISSY.add('dd-draggable', function(S) {
         _handleMouseDown: function(ev) {
             var self = this,
                 t = new S.Node(ev.target);
+
             if (!self._check(t)) return;
             //chrome 阻止了 flash 点击？？
             if (!UA.webkit) {