+++ src/ext/constrain-ext.js
@@ -1,3 +1,7 @@
+/**
+ * constrain extension for kissy
+ * @author:承玉<yiminghe@gmail.com>,乔花<qiaohua@taobao.com>
+ */
 KISSY.add("ext-constrain", function(S) {
     S.namespace("Ext");
 
@@ -26,7 +30,7 @@ KISSY.add("ext-constrain", function(S) {
      * @return {Object | undefined} {left: 0, top: 0, maxLeft: 100, maxTop: 100}
      */
     function _getConstrainRegion(constrain) {
-        var ret = undefined;
+        var ret;
         if (!constrain) return ret;
         var el = this.get("el");
         if (constrain !== true) {
@@ -39,13 +43,20 @@ KISSY.add("ext-constrain", function(S) {
         }
         // 没有指定 constrain, 表示受限于可视区域
         else {
+            //不要使用 viewportWidth()
+            //The innerWidth attribute, on getting,
+            //must return the viewport width including the size of a rendered scroll bar (if any).
+            //On getting, the clientWidth attribute returns the viewport width
+            //excluding the size of a rendered scroll bar (if any)
+            //  if the element is the root element 
+            var vWidth = document.documentElement.clientWidth;
             ret = { left: DOM.scrollLeft(), top: DOM.scrollTop() };
-
             S.mix(ret, {
-                maxLeft: ret.left + DOM.viewportWidth() - el[0].offsetWidth,
+                maxLeft: ret.left + vWidth - el[0].offsetWidth,
                 maxTop: ret.top + DOM.viewportHeight() - el[0].offsetHeight
             });
         }
+
         return ret;
     }
 
@@ -56,42 +67,42 @@ KISSY.add("ext-constrain", function(S) {
         },
         _renderUIConstrain:function() {
             S.log("_renderUIConstrain");
+            var self = this,
+                attrs = self.getDefAttrs(),
+                xAttr = attrs["x"],
+                yAttr = attrs["y"],
+                oriXSetter = xAttr["setter"],
+                oriYSetter = yAttr["setter"];
+            xAttr.setter = function(v) {
+                var r = oriXSetter && oriXSetter(v);
+                if (r === undefined) {
+                    r = v;
+                }
+                if (!self.get("constrain")) return r;
+                var _ConstrainExtRegion = _getConstrainRegion.call(
+                    self, self.get("constrain"));
+                return Math.min(Math.max(r,
+                    _ConstrainExtRegion.left),
+                    _ConstrainExtRegion.maxLeft);
+            };
+            yAttr.setter = function(v) {
+                var r = oriYSetter && oriYSetter(v);
+                if (r === undefined) {
+                    r = v;
+                }
+                if (!self.get("constrain")) return r;
+                var _ConstrainExtRegion = _getConstrainRegion.call(
+                    self, self.get("constrain"));
+                return Math.min(Math.max(r,
+                    _ConstrainExtRegion.top),
+                    _ConstrainExtRegion.maxTop);
+            };
+            self.addAttr("x", xAttr);
+            self.addAttr("y", yAttr);
+        },
 
-        },
-        _constrainX:function(ev) {
-            var self=this;
-            var v = ev.newVal,
-                _ConstrainExtRegion = _getConstrainRegion.call(
-                    self, self.get("constrain"));
-            if (!_ConstrainExtRegion) return;
-            if (v >= _ConstrainExtRegion.maxLeft
-                || v <= _ConstrainExtRegion.left) return false;
-        },
-        _constrainY:function(ev) {
-            var self=this;
-            var v = ev.newVal,
-                _ConstrainExtRegion = _getConstrainRegion.call(
-                    self, self.get("constrain"));
-            if (!_ConstrainExtRegion) return;
-            if (v >= _ConstrainExtRegion.maxTop
-                || v <= _ConstrainExtRegion.top) return false;
-        },
         _syncUIConstrain:function() {
             S.log("_syncUIConstrain");
-        },
-
-        _uiSetConstrain:function(v) {
-            S.log("_uiSetConstrain");
-            var self = this;
-            if (v) {
-                self.detach("beforeXChange", self._constrainX, self);
-                self.detach("beforeYChange", self._constrainY, self);
-                self.on("beforeXChange", self._constrainX, self);
-                self.on("beforeYChange", self._constrainY, self);
-            } else {
-                self.detach("beforeXChange", self._constrainX, self);
-                self.detach("beforeYChange", self._constrainY, self);
-            }
         },
         __destructor:function() {
             S.log("constrain-ext __destructor");