+++ src/ext/drag-ext.js
@@ -1,6 +1,6 @@
 /**
  * drag extension for position
- * @author:yiminghe@gmail.com
+ * @author:承玉<yiminghe@gmail.com>
  */
 KISSY.add("ext-drag", function(S) {
     S.namespace('Ext');
@@ -19,7 +19,7 @@ KISSY.add("ext-drag", function(S) {
 
     DragExt.prototype = {
 
-        _uiSetHanlders:function(v) {
+        _uiSetHandlers:function(v) {
             S.log("_uiSetHanlders");
             if (v && v.length > 0)
                 this.__drag.set("handlers", v);